+++ financius/src/main/java/com/code44/finance/modules/UIModule.java
@@ -19,9 +19,8 @@ import com.code44.finance.ui.dialogs.DeleteDialogFragment;
 import com.code44.finance.ui.dialogs.ListDialogFragment;
 import com.code44.finance.ui.dialogs.TimePickerDialog;
 import com.code44.finance.ui.overview.OverviewActivity;
-import com.code44.finance.ui.reports.CategoriesReportActivity;
-import com.code44.finance.ui.reports.CategoriesReportFragment;
 import com.code44.finance.ui.reports.ReportsFragment;
+import com.code44.finance.ui.reports.categories.CategoriesReportActivity;
 import com.code44.finance.ui.settings.SettingsActivity;
 import com.code44.finance.ui.settings.about.AboutActivity;
 import com.code44.finance.ui.settings.about.AboutFragment;
@@ -74,7 +73,6 @@ import dagger.Module;
                 GoogleApiFragment.class,
                 DataFragment.class,
                 ReportsFragment.class,
-                CategoriesReportFragment.class,
                 AboutFragment.class,
                 DatePickerDialog.class,
                 TimePickerDialog.class