+++ financius/src/main/java/com/code44/finance/ui/DrawerActivity.java
@@ -17,7 +17,7 @@ import com.code44.finance.adapters.NavigationAdapter;
 import com.code44.finance.ui.accounts.AccountsActivity;
 import com.code44.finance.ui.common.BaseActivity;
 import com.code44.finance.ui.overview.OverviewActivity;
-import com.code44.finance.ui.reports.CategoriesReportActivity;
+import com.code44.finance.ui.reports.categories.CategoriesReportActivity;
 import com.code44.finance.ui.transactions.TransactionsActivity;
 
 public abstract class DrawerActivity extends BaseActivity implements NavigationFragment.NavigationListener {