+++ financius/src/main/java/com/code44/finance/ui/reports/BaseReportActivity.java
@@ -1,0 +1,6 @@
+package com.code44.finance.ui.reports;
+
+import com.code44.finance.ui.DrawerActivity;
+
+public abstract class BaseReportActivity extends DrawerActivity {
+}